requirements.txt
pandas
numpy
python-binance
python-dotenv
pytz
requests
ezodf